+++ cmd/local_dev_setup/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"log"
 
-	"github.com/hex337/alex-koin-go/Models"
+	"github.com/hex337/alex-koin-go/Model"
 	"github.com/hex337/alex-koin-go/Config"
 	"github.com/davecgh/go-spew/spew"
 )
@@ -14,31 +14,31 @@ func main() {
 	Config.DBOpen()
 	// defer Config.DB.Close()
 
-	Config.DB.Migrator().DropTable(&Models.User{})
-	Config.DB.Migrator().DropTable(&Models.Coin{})
-	Config.DB.Migrator().DropTable(&Models.Transaction{})
+	Config.DB.Migrator().DropTable(&Model.User{})
+	Config.DB.Migrator().DropTable(&Model.Coin{})
+	Config.DB.Migrator().DropTable(&Model.Transaction{})
 
-	Config.DB.AutoMigrate(&Models.Transaction{}, &Models.User{}, &Models.Coin{})
+	Config.DB.AutoMigrate(&Model.Transaction{}, &Model.User{}, &Model.Coin{})
 
-	user1 := &Models.User{ FirstName: "Alex", LastName: "Koin" }
-	user2 := &Models.User{ FirstName: "Koin", LastName: "Lord" }
+	user1 := &Model.User{ FirstName: "Alex", LastName: "Koin" }
+	user2 := &Model.User{ FirstName: "Koin", LastName: "Lord" }
 
-	err =  Models.CreateUser(user1)
+	err =  Model.CreateUser(user1)
 	if err != nil {
 		log.Fatalf("Could not create user : %s", err.Error())
 	}
 
-	err =  Models.CreateUser(user2)
+	err =  Model.CreateUser(user2)
 	if err != nil {
 		log.Fatalf("Could not create user : %s", err.Error())
 	}
 
-	coin := &Models.Coin{ 
+	coin := &Model.Coin{ 
 		Origin: "Reason for the season",
 		MinedByUserID: user1.Model.ID,
 		UserID: user2.Model.ID,
 		CreatedByUserId: user1.Model.ID,
-		Transactions: []Models.Transaction{
+		Transactions: []Model.Transaction{
 			{
 				Amount: 1,
 				Memo: "Initial Koin Creation",
@@ -48,20 +48,20 @@ func main() {
 	  },
   }
 
-	err =  Models.CreateCoin(coin)
+	err =  Model.CreateCoin(coin)
 	if err != nil {
 		log.Fatalf("Could not create coin : %s", err.Error())
 	}
 
-	var user Models.User
-	err = Models.GetUserByID(&user, "1")
+	var user Model.User
+	err = Model.GetUserByID(&user, "1")
 	if err != nil {
 		log.Println(err.Error())
 	}
 
 	spew.Dump(user1)
 	log.Printf("\nUser 1 Balance : %d\nUser 2 Balance : %d", 
-		Models.GetUserBalance(user1),
-		Models.GetUserBalance(user2),
+		Model.GetUserBalance(user1),
+		Model.GetUserBalance(user2),
 	)
 }
